+++ src/EncryptionValidatorLight.sol
@@ -10,7 +10,7 @@ pragma solidity ^0.8.13;
 library EncryptionValidatorLight {
     /// @notice Minimum length required for encrypted data
     uint256 public constant MIN_ENCRYPTED_LENGTH = 40;
-    
+
     /**
      * @notice Quick check if a message appears encrypted
      * @dev Uses sampling to drastically reduce gas costs
@@ -19,22 +19,22 @@ library EncryptionValidatorLight {
      */
     function isEncryptedLight(string memory message) internal pure returns (bool isValid) {
         bytes memory data = bytes(message);
-        
+
         // 1. Check minimum length
         if (data.length < MIN_ENCRYPTED_LENGTH) {
             return false;
         }
-        
+
         // 2. Quick plain text check (only first bytes)
         if (isQuickPlainText(data)) {
             return false;
         }
-        
+
         // 3. Simplified entropy check with sampling
         if (!hasBasicEntropy(data)) {
             return false;
         }
-        
+
         return true;
     }
 
@@ -48,16 +48,16 @@ library EncryptionValidatorLight {
         uint256 printableCount = 0;
         uint256 spaceCount = 0;
         uint256 sampleSize = data.length > 32 ? 32 : data.length; // Only 32 bytes!
-        
-        for (uint i = 0; i < sampleSize; i++) {
+
+        for (uint256 i = 0; i < sampleSize; i++) {
             uint8 char = uint8(data[i]);
-            
+
             if (char >= 32 && char <= 126) {
                 printableCount++;
                 if (char == 32) spaceCount++; // space
             }
         }
-        
+
         // If >90% are readable characters AND there are spaces = probably text
         return (printableCount * 100 / sampleSize > 90) && (spaceCount > 0);
     }
@@ -70,12 +70,12 @@ library EncryptionValidatorLight {
      */
     function hasBasicEntropy(bytes memory data) internal pure returns (bool) {
         if (data.length == 0) return false;
-        
+
         // For small data, check all bytes
         if (data.length <= 64) {
             return hasFullEntropy(data);
         }
-        
+
         // For large data, use sampling of only 64 bytes
         return hasSampledEntropy(data, 64);
     }
@@ -85,18 +85,18 @@ library EncryptionValidatorLight {
      */
     function hasFullEntropy(bytes memory data) internal pure returns (bool) {
         uint256[256] memory frequency;
-        
-        for (uint i = 0; i < data.length; i++) {
+
+        for (uint256 i = 0; i < data.length; i++) {
             frequency[uint8(data[i])]++;
         }
-        
+
         uint256 uniqueValues = 0;
-        for (uint i = 0; i < 256; i++) {
+        for (uint256 i = 0; i < 256; i++) {
             if (frequency[i] > 0) {
                 uniqueValues++;
             }
         }
-        
+
         // For small data, needs at least 60% unique bytes
         return (uniqueValues * 100 / data.length) >= 60;
     }
@@ -106,20 +106,20 @@ library EncryptionValidatorLight {
      */
     function hasSampledEntropy(bytes memory data, uint256 sampleSize) internal pure returns (bool) {
         uint256[256] memory frequency;
-        
+
         // Uniformly distributed sampling
-        for (uint i = 0; i < sampleSize; i++) {
+        for (uint256 i = 0; i < sampleSize; i++) {
             uint256 index = (i * data.length) / sampleSize;
             frequency[uint8(data[index])]++;
         }
-        
+
         uint256 uniqueValues = 0;
-        for (uint i = 0; i < 256; i++) {
+        for (uint256 i = 0; i < 256; i++) {
             if (frequency[i] > 0) {
                 uniqueValues++;
             }
         }
-        
+
         // For sampled data, expects at least 50% unique values
         return (uniqueValues * 100 / sampleSize) >= 50;
     }
@@ -127,24 +127,21 @@ library EncryptionValidatorLight {
     /**
      * @notice Test with detailed feedback (light version)
      */
-    function testEncryptionLight(string memory message) internal pure returns (
-        bool isValid,
-        string memory reason
-    ) {
+    function testEncryptionLight(string memory message) internal pure returns (bool isValid, string memory reason) {
         bytes memory data = bytes(message);
-        
+
         if (data.length < MIN_ENCRYPTED_LENGTH) {
             return (false, "Message too short");
         }
-        
+
         if (isQuickPlainText(data)) {
             return (false, "Looks like plain text");
         }
-        
+
         if (!hasBasicEntropy(data)) {
             return (false, "Low entropy - doesn't look encrypted");
         }
-        
+
         return (true, "Looks encrypted");
     }
 }
